+++ packages/orq-rc/src/models/operations/createconversationresponse.ts
@@ -337,7 +337,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kfg0379v46kd78p305h27zg4"),
+  _id: z.string().default("reasoning_01kfg3pf5he6316f6drfkb5211"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -362,7 +362,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kfg0379v46kd78p305h27zg4"),
+  id: z.string().default("reasoning_01kfg3pf5he6316f6drfkb5211"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),
@@ -510,7 +510,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kfg0379meegwa9t23tne9pmj"),
+  _id: z.string().default("reasoning_01kfg3pf5amank95k18hgfnxfx"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$inboundSchema,
   reasoning: z.string(),
@@ -535,7 +535,7 @@ export const ReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kfg0379meegwa9t23tne9pmj"),
+  id: z.string().default("reasoning_01kfg3pf5amank95k18hgfnxfx"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$outboundSchema,
   reasoning: z.string(),